+++ history.md
@@ -1,3 +1,7 @@
+# 2022-02-23
+
+- Add workspace.openTextDocument()
+
 # 2022-02-18
 
 - Clear all highlights created by coc.nvim before restart.